requirements.txt
selenium>=4.0.0,<5
webdriver-manager>=1.8.2,<2
jsonschema>=3.0.1,<4
pytest>=6.2.5, <7
allure-pytest
pydash>=4.7.5,<5
openpyxl
pytest-xdist>=1.29.0,<2
allure-pytest
allure-python-commons
pytest-xdist>=1.29.0,<2
retrying>=1.3.3,<2
xlsxwriter
pytest-rerunfailures>=10.2